refactor: rename test files for consistency
- Renamed test_errors.py -> test_media_errors.py
- Renamed test_performance.py -> test_media_performance.py
- Renamed test_sync/test_errors.py -> test_sync/test_sync_errors.py
- Renamed test_sync/test_performance.py -> test_sync/test_sync_performance.py

This makes the test file naming more consistent by:
1. Including the module name in the test file name
2. Making it clear which tests belong to which module
3. Following the pattern: test_<module>_<feature>.py
